Add skip of log() methods
This adds skipping of all methods that are named "log", for displaying the
locations that call tr.Ace.log. With this change, a class can have a single
"log" method, possibly filtering for displaying the logging statements from
that class (and its subclasses).

For example (pseudo-code):

    class MyClass {
         boolean debug = true;

         void log(String msg, Object obj) {
             if (debug) {
                 tr.Ace.log(msg, obj);
             }
         }

         void start(Object value) {
              log("initializing", value);
         }

         void end(Object value) {
              log("terminating", value);
         }
     }

This will display (again, pseudo-output):

    [MyClass     14] {MyClass      #start    } initializing: x
    [            18] {             #end      } terminating: y